+++ src/functions/index.ts
@@ -2,3 +2,4 @@ export * from "./did";
 export * from "./unik";
 export * from "./certifications";
 export * from "./jwt";
+export * from "./transactions";